requirements.txt
google-genai
pillow
streamlit
gspread
